Prettify fn names
